+++ design/src/main/kotlin/uk/govuk/app/design/ui/component/Header.kt
@@ -13,50 +13,12 @@ import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.focus.FocusRequester
-import androidx.compose.ui.focus.focusRequester
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import uk.govuk.app.design.R
 import uk.govuk.app.design.ui.theme.GovUkTheme
-
-@Composable
-fun SearchHeader(
-    onBack: () -> Unit,
-    placeholder: String,
-    onSearch: (String) -> Unit,
-    onClear: () -> Unit,
-    modifier: Modifier = Modifier,
-    focusRequester: FocusRequester = FocusRequester()
-) {
-    Column(modifier.fillMaxWidth()) {
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            TextButton(
-                onClick = onBack
-            ) {
-                Icon(
-                    imageVector = Icons.AutoMirrored.Filled.ArrowBack,
-                    contentDescription = stringResource(R.string.content_desc_back),
-                    tint = GovUkTheme.colourScheme.textAndIcons.link
-                )
-            }
-            SearchField(
-                placeholder = placeholder,
-                onSearch = onSearch,
-                onClear = onClear,
-                modifier = Modifier
-                    .weight(1f)
-                    .focusRequester(focusRequester)
-            )
-        }
-        ListDivider()
-    }
-}
 
 @Composable
 fun TabHeader(
@@ -104,19 +66,6 @@ fun ChildPageHeader(
             modifier = modifier
                 .fillMaxWidth()
                 .padding(horizontal = GovUkTheme.spacing.medium)
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun SearchHeaderPreview() {
-    GovUkTheme {
-        SearchHeader(
-            onBack = { },
-            placeholder = "Search",
-            onSearch = { },
-            onClear = { }
         )
     }
 }
